+++ packages/nfts/script/trailblazers-season-2/Deploy.s.sol
@@ -22,11 +22,12 @@ contract DeployS2Script is Script {
     BadgeRecruitment recruitment;
 
     // Taiko Mainnet Values
+    /*
     //address owner = 0xf8ff2AF0DC1D5BA4811f22aCb02936A1529fd2Be;
     address claimMintSigner = 0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266;
     address recruitmentSigner = 0x9Fc8d56c7376f9b062FEe7E02BAdFA670d603248;
     string baseURI =
-        "https://taikonfts.4everland.link/ipfs/bafybeiatuzeeeznd3hi5qiulslxcjd22ebu45t4fra2jvi3smhocr2c66a";
+    "https://taikonfts.4everland.link/ipfs/bafybeiatuzeeeznd3hi5qiulslxcjd22ebu45t4fra2jvi3smhocr2c66a";
     IMinimalBlacklist blacklist = IMinimalBlacklist(0xfA5EA6f9A13532cd64e805996a941F101CCaAc9a);
 
     uint256 public MAX_INFLUENCES = 5;
@@ -37,11 +38,10 @@ contract DeployS2Script is Script {
     uint256 public DEFAULT_CYCLE_DURATION = 7 days;
     uint256 public s1EndDate = 1_734_350_400; // Dec 16th 2024, noon UTC
     uint256 public S1_LOCK_DURATION = (s1EndDate - block.timestamp);
-
+    */
     // Hekla Testnet Values
-    /*
     string baseURI =
-    "https://taikonfts.4everland.link/ipfs/bafybeiatuzeeeznd3hi5qiulslxcjd22ebu45t4fra2jvi3smhocr2c66a";
+        "https://taikonfts.4everland.link/ipfs/bafybeiatuzeeeznd3hi5qiulslxcjd22ebu45t4fra2jvi3smhocr2c66a";
 
     IMinimalBlacklist blacklist = IMinimalBlacklist(0xe61E9034b5633977eC98E302b33e321e8140F105);
     address claimMintSigner = 0xf39Fd6e51aad88F6F4ce6aB8827279cffFb92266;
@@ -54,7 +54,6 @@ contract DeployS2Script is Script {
     uint256 public MAX_INFLUENCES_DIVIDER = 100;
     uint256 public DEFAULT_CYCLE_DURATION = 7 days;
     uint256 public S1_LOCK_DURATION = 365 days;
-    */
 
     address s1Contract = 0xa20a8856e00F5ad024a55A663F06DCc419FFc4d5;