+++ packages/taiko-client/driver/chain_syncer/blob/syncer.go
@@ -207,6 +207,7 @@ func (s *Syncer) onBlockProposed(
 	}
 	// Ignore those already inserted blocks.
 	if s.lastInsertedBlockID != nil && meta.GetBlockID().Cmp(s.lastInsertedBlockID) <= 0 {
+		log.Debug("Skip already inserted block", "blockID", meta.GetBlockID(), "lastInsertedBlockID", s.lastInsertedBlockID)
 		return nil
 	}
 
@@ -233,6 +234,7 @@ func (s *Syncer) onBlockProposed(
 	if s.progressTracker.Triggered() {
 		// Already synced through beacon sync, just skip this event.
 		if meta.GetBlockID().Cmp(s.progressTracker.LastSyncedBlockID()) <= 0 {
+			log.Debug("Skip already beacon synced block", "blockID", meta.GetBlockID())
 			return nil
 		}
 
@@ -424,7 +426,7 @@ func (s *Syncer) insertNewHead(
 
 	fc := &engine.ForkchoiceStateV1{
 		HeadBlockHash:      payload.BlockHash,
-		SafeBlockHash:      payload.BlockHash,
+		SafeBlockHash:      lastVerifiedBlockHash,
 		FinalizedBlockHash: lastVerifiedBlockHash,
 	}