+++ packages/taiko-client/driver/chain_syncer/blob/syncer_test.go
@@ -11,10 +11,13 @@ import (
 	"golang.org/x/exp/slog"
 
 	"github.com/ethereum-optimism/optimism/op-service/txmgr"
+	"github.com/ethereum/go-ethereum/beacon/engine"
 	"github.com/ethereum/go-ethereum/common"
+	"github.com/ethereum/go-ethereum/consensus/taiko"
 	"github.com/ethereum/go-ethereum/core/rawdb"
 	"github.com/ethereum/go-ethereum/core/types"
 	"github.com/ethereum/go-ethereum/crypto"
+	"github.com/ethereum/go-ethereum/rlp"
 	"github.com/stretchr/testify/suite"
 
 	"github.com/taikoxyz/taiko-mono/packages/taiko-client/bindings"
@@ -57,6 +60,99 @@ func (s *BlobSyncerTestSuite) SetupTest() {
 	s.s = syncer
 
 	s.initProposer()
+}
+
+func (s *BlobSyncerTestSuite) TestBlobSyncRobustness() {
+	ctx := context.Background()
+
+	meta := s.ProposeAndInsertValidBlock(s.p, s.s)
+
+	block, err := s.RPCClient.L2.BlockByNumber(ctx, meta.GetBlockID())
+	s.Nil(err)
+
+	lastVerifiedBlockInfo, err := s.s.rpc.GetLastVerifiedBlock(ctx)
+	s.Nil(err)
+
+	txListBytes, err := rlp.EncodeToBytes(block.Transactions())
+	s.Nil(err)
+
+	parent, err := s.RPCClient.L2ParentByBlockID(context.Background(), meta.GetBlockID())
+	s.Nil(err)
+
+	// Reset l2 chain.
+	s.Nil(rpc.SetHead(ctx, s.RPCClient.L2, common.Big0))
+
+	attributes := &engine.PayloadAttributes{
+		Timestamp:             meta.GetTimestamp(),
+		Random:                meta.GetDifficulty(),
+		SuggestedFeeRecipient: meta.GetCoinbase(),
+		Withdrawals:           make([]*types.Withdrawal, 0),
+		BlockMetadata: &engine.BlockMetadata{
+			Beneficiary: meta.GetCoinbase(),
+			GasLimit:    uint64(meta.GetGasLimit()) + taiko.AnchorGasLimit,
+			Timestamp:   meta.GetTimestamp(),
+			TxList:      txListBytes,
+			MixHash:     meta.GetDifficulty(),
+			ExtraData:   meta.GetExtraData(),
+		},
+		BaseFeePerGas: block.BaseFee(),
+		L1Origin: &rawdb.L1Origin{
+			BlockID:       meta.GetBlockID(),
+			L2BlockHash:   common.Hash{}, // Will be set by taiko-geth.
+			L1BlockHeight: meta.GetRawBlockHeight(),
+			L1BlockHash:   meta.GetRawBlockHash(),
+		},
+	}
+
+	step0 := func() *engine.ForkChoiceResponse {
+		fcRes, err := s.RPCClient.L2Engine.ForkchoiceUpdate(
+			ctx,
+			&engine.ForkchoiceStateV1{HeadBlockHash: parent.Hash()},
+			attributes,
+		)
+		s.Nil(err)
+		s.Equal(engine.VALID, fcRes.PayloadStatus.Status)
+		s.True(true, fcRes.PayloadID != nil)
+		return fcRes
+	}
+
+	step1 := func(fcRes *engine.ForkChoiceResponse) *engine.ExecutableData {
+		payload, err := s.RPCClient.L2Engine.GetPayload(ctx, fcRes.PayloadID)
+		s.Nil(err)
+		return payload
+	}
+
+	step2 := func(payload *engine.ExecutableData) *engine.ExecutableData {
+		execStatus, err := s.RPCClient.L2Engine.NewPayload(ctx, payload)
+		s.Nil(err)
+		s.Equal(engine.VALID, execStatus.Status)
+		return payload
+	}
+
+	step3 := func(payload *engine.ExecutableData) {
+		fcRes, err := s.RPCClient.L2Engine.ForkchoiceUpdate(ctx, &engine.ForkchoiceStateV1{
+			HeadBlockHash:      payload.BlockHash,
+			SafeBlockHash:      lastVerifiedBlockInfo.BlockHash,
+			FinalizedBlockHash: lastVerifiedBlockInfo.BlockHash,
+		}, nil)
+		s.Nil(err)
+		s.Equal(engine.VALID, fcRes.PayloadStatus.Status)
+	}
+
+	loopSize := 10
+	for i := 0; i < loopSize; i++ {
+		step0()
+	}
+
+	for i := 0; i < loopSize; i++ {
+		step1(step0())
+	}
+
+	for i := 0; i < loopSize; i++ {
+		step2(step1(step0()))
+	}
+
+	step3(step2(step1(step0())))
 }
 
 func (s *BlobSyncerTestSuite) TestProcessL1Blocks() {