+++ packages/taiko-client/pkg/chain_iterator/event_iterator/block_proposed_iterator.go
@@ -3,10 +3,12 @@ package eventiterator
 import (
 	"context"
 	"errors"
+	"fmt"
 	"math/big"
 
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
 	"github.com/ethereum/go-ethereum/core/types"
+	"github.com/ethereum/go-ethereum/log"
 
 	"github.com/taikoxyz/taiko-mono/packages/taiko-client/bindings"
 	"github.com/taikoxyz/taiko-mono/packages/taiko-client/bindings/metadata"
@@ -109,17 +111,12 @@ func assembleBlockProposedIteratorCallback(
 		updateCurrentFunc chainIterator.UpdateCurrentFunc,
 		endFunc chainIterator.EndIterFunc,
 	) error {
-		endHeight := end.Number.Uint64()
-
-		iter, err := taikoL1.FilterBlockProposed(
-			&bind.FilterOpts{Start: start.Number.Uint64(), End: &endHeight, Context: ctx},
-			filterQuery,
-			nil,
+		var (
+			endHeight   = end.Number.Uint64()
+			lastBlockID uint64
 		)
-		if err != nil {
-			return err
-		}
-		defer iter.Close()
+
+		log.Debug("Iterating BlockProposed events", "start", start.Number, "end", endHeight)
 
 		iterOntake, err := taikoL1.FilterBlockProposedV2(
 			&bind.FilterOpts{Start: start.Number.Uint64(), End: &endHeight, Context: ctx},
@@ -130,14 +127,31 @@ func assembleBlockProposedIteratorCallback(
 		}
 		defer iterOntake.Close()
 
-		for iter.Next() {
-			event := iter.Event
+		for iterOntake.Next() {
+			event := iterOntake.Event
+			log.Debug("Processing BlockProposedV2 event", "block", event.BlockId, "l1BlockHeight", event.Raw.BlockNumber)
 
-			if err := callback(ctx, metadata.NewTaikoDataBlockMetadataLegacy(event), eventIter.end); err != nil {
+			if lastBlockID != 0 && event.BlockId.Uint64() != lastBlockID+1 {
+				log.Warn(
+					"BlockProposedV2 event is not continuous, rescan the L1 chain",
+					"fromL1Block", start.Number,
+					"toL1Block", endHeight,
+					"lastScannedBlockID", lastBlockID,
+					"currentScannedBlockID", event.BlockId.Uint64(),
+				)
+				return fmt.Errorf(
+					"BlockProposedV2 event is not continuous, lastScannedBlockID: %d, currentScannedBlockID: %d",
+					lastBlockID, event.BlockId.Uint64(),
+				)
+			}
+
+			if err := callback(ctx, metadata.NewTaikoDataBlockMetadataOntake(event), eventIter.end); err != nil {
+				log.Warn("Error while processing BlockProposedV2 events, keep retrying", "error", err)
 				return err
 			}
 
 			if eventIter.isEnd {
+				log.Debug("BlockProposedIterator is ended", "start", start.Number, "end", endHeight)
 				endFunc()
 				return nil
 			}
@@ -146,28 +160,17 @@ func assembleBlockProposedIteratorCallback(
 			if err != nil {
 				return err
 			}
+
+			log.Debug("Updating current block cursor for processing BlockProposedV2 events", "block", current.Number)
+
+			lastBlockID = event.BlockId.Uint64()
 
 			updateCurrentFunc(current)
 		}
 
-		for iterOntake.Next() {
-			event := iterOntake.Event
-
-			if err := callback(ctx, metadata.NewTaikoDataBlockMetadataOntake(event), eventIter.end); err != nil {
-				return err
-			}
-
-			if eventIter.isEnd {
-				endFunc()
-				return nil
-			}
-
-			current, err := client.HeaderByHash(ctx, event.Raw.BlockHash)
-			if err != nil {
-				return err
-			}
-
-			updateCurrentFunc(current)
+		if err := iterOntake.Error(); err != nil {
+			log.Error("Error while iterating BlockProposedV2 events", "error", err)
+			return err
 		}
 
 		return nil