+++ packages/taiko-client/pkg/rpc/methods.go
@@ -76,6 +76,9 @@ func (c *Client) ensureGenesisMatched(ctx context.Context) error {
 		if iter.Next() {
 			l2GenesisHash = iter.Event.BlockHash
 		}
+		if iter.Error() != nil {
+			return iter.Error()
+		}
 	} else {
 		// Fetch the genesis `BlockVerified` event.
 		iter, err := c.TaikoL1.FilterBlockVerified(filterOpts, []*big.Int{common.Big0}, nil)
@@ -85,6 +88,9 @@ func (c *Client) ensureGenesisMatched(ctx context.Context) error {
 
 		if iter.Next() {
 			l2GenesisHash = iter.Event.BlockHash
+		}
+		if iter.Error() != nil {
+			return iter.Error()
 		}
 	}
 
@@ -334,50 +340,23 @@ func (c *Client) GetPoolContent(
 	ctxWithTimeout, cancel := CtxWithTimeoutOrDefault(ctx, defaultTimeout)
 	defer cancel()
 
-	// Get the latest L2 block header at first.
+	l1Head, err := c.L1.HeaderByNumber(ctx, nil)
+	if err != nil {
+		return nil, err
+	}
+
 	l2Head, err := c.L2.HeaderByNumber(ctx, nil)
 	if err != nil {
 		return nil, err
-	}
-
-	l1Origin, err := c.L2.L1OriginByID(ctx, l2Head.Number)
-	if err != nil && err.Error() != ethereum.NotFound.Error() {
-		return nil, err
-	}
-
-	var (
-		L1HeadNum *big.Int
-		L2HeadNum *big.Int
-		timestamp = uint64(time.Now().Unix())
-	)
-
-	if l1Origin != nil && l1Origin.IsSoftBlock() && !l1Origin.EndOfPreconf && !l1Origin.EndOfBlock {
-		// Check if this is an unfinished soft block, if not, we will use the latest L1 / L2 block number from the L1Origin.
-		// Otherwise, we will use the L1 / L2 block number in L1Origin.
-		L1HeadNum = l1Origin.L1BlockHeight
-		L2HeadNum = new(big.Int).Sub(l1Origin.BlockID, common.Big1)
-	}
-
-	l1Head, err := c.L1.HeaderByNumber(ctx, L1HeadNum)
-	if err != nil {
-		return nil, err
-	}
-
-	if L2HeadNum != nil {
-		timestamp = l2Head.Time
-		l2Head, err = c.L2.HeaderByNumber(ctx, L2HeadNum)
-		if err != nil {
-			return nil, err
-		}
 	}
 
 	baseFee, err := c.CalculateBaseFee(
 		ctx,
 		l2Head,
 		l1Head.Number,
-		true,
+		chainConfig.IsOntake(new(big.Int).Add(l2Head.Number, common.Big1)),
 		&chainConfig.ProtocolConfigs.BaseFeeConfig,
-		timestamp,
+		uint64(time.Now().Unix()),
 	)
 	if err != nil {
 		return nil, err
@@ -674,10 +653,12 @@ func (c *Client) checkSyncedL1SnippetFromAnchor(
 	log.Debug("Check synced L1 snippet from anchor", "blockID", blockID, "l1Height", l1Height)
 	block, err := c.L2.BlockByNumber(ctx, blockID)
 	if err != nil {
+		log.Error("Failed to fetch L2 block", "blockID", blockID, "error", err)
 		return false, err
 	}
 	parent, err := c.L2.BlockByHash(ctx, block.ParentHash())
 	if err != nil {
+		log.Error("Failed to fetch L2 parent block", "blockID", blockID, "parentHash", block.ParentHash(), "error", err)
 		return false, err
 	}
 
@@ -685,6 +666,7 @@ func (c *Client) checkSyncedL1SnippetFromAnchor(
 		block.Transactions()[0],
 	)
 	if err != nil {
+		log.Error("Failed to parse L1 snippet from anchor transaction", "blockID", blockID, "error", err)
 		return false, err
 	}
 
@@ -700,6 +682,7 @@ func (c *Client) checkSyncedL1SnippetFromAnchor(
 
 	l1Header, err := c.L1.HeaderByNumber(ctx, new(big.Int).SetUint64(l1HeightInAnchor))
 	if err != nil {
+		log.Error("Failed to fetch L1 header", "blockID", blockID, "error", err)
 		return false, err
 	}
 
@@ -727,7 +710,7 @@ func (c *Client) getSyncedL1SnippetFromAnchor(
 ) {
 	method, err := encoding.TaikoL2ABI.MethodById(tx.Data())
 	if err != nil {
-		return common.Hash{}, 0, 0, err
+		return common.Hash{}, 0, 0, fmt.Errorf("failed to get TaikoL2.Anchor method by ID: %w", err)
 	}
 
 	var ok bool
@@ -736,7 +719,7 @@ func (c *Client) getSyncedL1SnippetFromAnchor(
 		args := map[string]interface{}{}
 
 		if err := method.Inputs.UnpackIntoMap(args, tx.Data()[4:]); err != nil {
-			return common.Hash{}, 0, 0, err
+			return common.Hash{}, 0, 0, fmt.Errorf("failed to unpack anchor transaction calldata: %w", err)
 		}
 
 		l1StateRoot, ok = args["_l1StateRoot"].([32]byte)
@@ -860,6 +843,9 @@ func (c *Client) GetTaikoDataSlotBByNumber(ctx context.Context, number uint64) (
 
 	for iter.Next() {
 		return &iter.Event.SlotB, nil
+	}
+	if iter.Error() != nil {
+		return nil, fmt.Errorf("failed to get state variables by block number %d: %w", number, iter.Error())
 	}
 
 	return nil, fmt.Errorf("failed to get state variables by block number %d", number)